requirements.txt
asgiref==3.8.1
autoflake==2.3.1
black==24.4.2
cfgv==3.4.0
click==8.1.7
coverage==7.5.1
distlib==0.3.8
Django==5.0.4
django-allauth==0.62.1
django-stubs==5.0.0
django-stubs-ext==5.0.0
filelock==3.14.0
identify==2.5.36
isort==5.13.2
mypy==1.10.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
pre-commit-hooks==4.6.0
pyflakes==3.2.0
PyYAML==6.0.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.4.4
setuptools==69.5.1
sqlparse==0.5.0
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
virtualenv==20.26.2
